// repos/sbs-draft-web-main/utils/types/types.ts
export type RosterProps = {
    playerId: string
    playerStateInfo: {
        displayName: string
        ownerAddress: string
        pickNum: number
        playerId: string
        position: string
        round: number
        team: string
    }
    stats: {
        averageScore: number
        highestScore: number
        playerId: string
        top5Finishes: number
        byeWeek: number
        adp: number
        playersFromTeam: null | string[]
    }
}
// ...
export type DraftRosterProps = {
    [key: string]: {
        QB: RosterProps[]
        RB: RosterProps[]
        WR: RosterProps[]
        TE: RosterProps[]
        DST: RosterProps[]
        PFP: {
            imageUrl: string
            nftContract: string
            displayName: string
        }
    }
}